feat: created ui of bookdetail and fetched data

+++ lib/ui/views/bookdetail/model/book_model.dart
@@ -313,9 +313,9 @@ class Paperback {
   dynamic width;
   dynamic height;
   dynamic thickness;
-  int? weight;
+  double? weight;
   int? mrp;
-  int? sellingPrice;
+  double? sellingPrice;
   int? productId;
   String? origin;
   String? isbn10;
